@@ -1,15 +1,17 @@
 {
-  "name": "4dat",
+  "name": "4dat-backend",
   "version": "1.0.0",
   "description": "proof of concept web app that will allow users to track their tools",
   "main": "app.js",
   "scripts": {
-    "start": "node ./node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-    "webpack": "./node_modules/.bin/webpack",
+    "start": "./node_modules/.bin/babel-node api/app.js --presets es2015,stage-2 && react-scripts start",
+    "dev": "nodemon api/app.js --exec babel-node --presets es2015,stage-2",
     "test": "mocha",
     "test:watch": "npm run test -- --watch",
-    "prod": "NODE_ENV=production node server.js",
-    "postinstall": "webpack -p . --history-api-fallback"
+    "test-client": "react-scripts test --env=jsdom",
+    "build": "react-scripts build",
+    "eject": "react-scripts eject",
+    "start-client": "react-scripts start"
   },
   "repository": {
     "type": "git",
@@ -20,80 +22,32 @@
   "bugs": {
     "url": "https://github.com/udacity-blitz/4dat/issues"
   },
-  "homepage": "https://github.com/udacity-blitz/4dat#readme",
   "dependencies": {
-    "@google-cloud/datastore": "^1.0.0",
-    "auth0-js": "^8.7.0",
-    "aws-lib": "^0.3.0",
-    "axios": "^0.16.1",
-    "babel-core": "^6.2.1",
-    "babel-loader": "^6.2.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-react": "^6.1.18",
-    "babel-preset-stage-1": "^6.1.18",
+    "apac": "^3.0.2",
+    "auth0-js": "^8.9.1",
+    "axios": "^0.16.2",
+    "babel-cli": "^6.24.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
     "body-parser": "^1.17.2",
     "bootstrap": "^3.3.7",
-    "cors": "^2.8.3",
-    "eslint": "^3.19.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-react": "^6.10.3",
-    "express": "^4.15.2",
+    "child_process": "^1.0.2",
+    "cors": "^2.8.4",
+    "express": "^4.15.4",
     "express-jwt": "^5.3.0",
-    "jwks-rsa": "^1.1.1",
+    "gstore-api": "^1.1.0",
+    "jwks-rsa": "^1.2.0",
     "jwt-decode": "^2.2.0",
-    "mocha": "^3.3.0",
-    "mongoose": "^4.9.6",
     "nconf": "^0.8.4",
-    "passport": "^0.3.2",
-    "passport-auth0": "^0.6.0",
-    "react": "^0.14.3",
-    "react-bootstrap": "^0.31.0",
-    "react-burger-menu": "^2.0.2",
-    "react-dom": "^0.14.3",
+    "path": "^0.12.7",
+    "react": "^15.6.1",
+    "react-bootstrap": "^0.31.2",
+    "react-dom": "^15.6.1",
     "react-icons": "^2.2.5",
-    "react-redux": "4.0.0",
-    "react-router": "^2.0.1",
-    "redux": "^3.0.4",
-    "redux-promise": "^0.5.3",
-    "standard": "^10.0.2",
-    "webpack": "^1.15.0",
-    "webpack-cluster": "^3.0.0",
-    "webpack-dev-server": "^1.16.5"
+    "react-router": "^2.8.1"
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "chai": "^3.5.0",
-    "css-loader": "^0.28.2",
-    "eslint": "^3.19.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-react": "^6.10.3",
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
-    "gulp-clean": "^0.3.1",
-    "gulp-clean-css": "^3.0.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-extname": "^0.2.0",
-    "gulp-htmlmin": "^3.0.0",
-    "gulp-load-plugins": "^0.7.0",
-    "gulp-rev": "^2.0.1",
-    "gulp-sass": "^3.1.0",
-    "gulp-template": "^1.1.1",
-    "gulp-uglify": "^2.0.1",
-    "gulp-util": "^3.0.8",
-    "gulp-watch": "^4.3.11",
-    "mocha": "^3.3.0",
-    "react-scripts": "0.9.5",
-    "sinon": "^2.1.0",
-    "standard": "^10.0.2",
-    "style-loader": "^0.18.1",
-    "webpack": "^2.4.1"
+    "mocha": "^3.5.0",
+    "react-scripts": "1.0.11"
   }
 }
