@@ -1,22 +1,7 @@
-// Copyright 2017, Google, Inc.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 'use strict';
 
-// Hierarchical node.js configuration with command-line arguments, environment
-// variables, and files.
-const nconf = module.exports = require('nconf');
-const path = require('path');
+import nconf from 'nconf';
+import path from 'path';
 
 nconf
   // 1. Command-line arguments
@@ -26,11 +11,6 @@ nconf
     'DATA_BACKEND',
     'GOOGLE_APPLICATION_CREDENTIALS',
     'GCLOUD_PROJECT',
-    'INSTANCE_CONNECTION_NAME',
-    'MONGO_URL',
-    'MONGO_COLLECTION',
-    'MYSQL_USER',
-    'MYSQL_PASSWORD',
     'NODE_ENV',
     'PORT'
   ])
@@ -38,42 +18,20 @@ nconf
   .file({ file: path.join(__dirname, 'config.json') })
   // 4. Defaults
   .defaults({
-    // dataBackend can be 'datastore', 'cloudsql', or 'mongodb'. Be sure to
-    // configure the appropriate settings for each storage engine below.
-    // If you are unsure, use datastore as it requires no additional
-    // configuration.
     DATA_BACKEND: 'datastore',
-    GOOGLE_APPLICATION_CREDENTIALS:'./customer-4dat-ccc5c3a860fb.json',
-    // This is the id of your project in the Google Cloud Developers Console.
     GCLOUD_PROJECT: 'customer-4dat',
-
-    // MongoDB connection string
-    // https://docs.mongodb.org/manual/reference/connection-string/
-    MONGO_URL: 'mongodb://localhost:27017',
-    MONGO_COLLECTION: 'books',
-
-    MYSQL_USER: '',
-    MYSQL_PASSWORD: '',
-
     PORT: 80
   });
 
 // Check for required settings
 checkConfig('GCLOUD_PROJECT');
-
-if (nconf.get('DATA_BACKEND') === 'cloudsql') {
-  checkConfig('MYSQL_USER');
-  checkConfig('MYSQL_PASSWORD');
-  if (nconf.get('NODE_ENV') === 'production') {
-    checkConfig('INSTANCE_CONNECTION_NAME');
-  }
-} else if (nconf.get('DATA_BACKEND') === 'mongodb') {
-  checkConfig('MONGO_URL');
-  checkConfig('MONGO_COLLECTION');
-}
+checkConfig('PORT');
+checkConfig('GOOGLE_APPLICATION_CREDENTIALS');
 
 function checkConfig (setting) {
   if (!nconf.get(setting)) {
     throw new Error(`You must set ${setting} as an environment variable or in config.json!`);
   }
 }
+
+export { nconf };
