@@ -2,7 +2,7 @@ import React from 'react';
 import '../css/Countbox.css';
 
 const Countbox = (props) => (
-  <div className={'box ' + props.itemColor }>
+  <div className={'box ' + props.itemColor } onClick={props.onClick}>
     <p className='title'> {props.itemName} </p>
     <p className='count'> {props.itemCount} </p>
   </div>
